+++ src/gst_image_view/GstImageView.h
@@ -69,7 +69,7 @@ public slots:
     void addCircle(QPointF &center, double radius, QColor &color, double width, bool persistent = 0);
     void addLine(QLineF &line, QColor &color, double width, bool persistent = 0);
     void addPolygon(QPolygonF &polygon, QColor &color, double width, bool persistent = 0);
-    void addPoints(const QList<int> points_x,QList<int> points_y, QColor &color,double width, bool persistent=0);
+    void addPoints(const QList<int> points_x,QList<int> points_y, QColor &color, int width, bool persistent=0);
 
     /** Writes text on top of the image. The text is immune to image transformation, i.e. rotation or scaling. */
     void addText(QString text, /*TextLocation*/ int location, QColor color = QColor(Qt::black), bool persistent = 0);
